+++ lib/models/player.dart
@@ -4,14 +4,16 @@ class Player {
   String age;
   String position;
   String team;
+  String teamName;
   String league;
 
-  Player(String id, String name, String age, String position, String team, String league) {
+  Player(String id, String name, String age, String position, String team, String teamName, String league) {
     this.id = id;
     this.name = name;
     this.age = age;
     this.position = position;
     this.team = team;
+    this.teamName = teamName;
     this.league = league;
   }
 }